refactor: Move irrelevant functions from the regex helper
