+++ lib/util/regex.ts
@@ -72,42 +72,3 @@ export function escapeRegExp(input: string): string {
 }
 
 export const newlineRegex = regEx(/\r?\n/);
-
-const configValStart = regEx(/^!?\//);
-const configValEnd = regEx(/\/i?$/);
-
-export function isConfigRegex(input: unknown): input is string {
-  return (
-    is.string(input) && configValStart.test(input) && configValEnd.test(input)
-  );
-}
-
-function parseConfigRegex(input: string): RegExp | null {
-  try {
-    const regexString = input
-      .replace(configValStart, '')
-      .replace(configValEnd, '');
-    return input.endsWith('i') ? regEx(regexString, 'i') : regEx(regexString);
-  } catch (err) {
-    // no-op
-  }
-  return null;
-}
-
-type ConfigRegexPredicate = (s: string) => boolean;
-
-export function configRegexPredicate(
-  input: string,
-): ConfigRegexPredicate | null {
-  if (isConfigRegex(input)) {
-    const configRegex = parseConfigRegex(input);
-    if (configRegex) {
-      const isPositive = !input.startsWith('!');
-      return (x: string): boolean => {
-        const res = configRegex.test(x);
-        return isPositive ? res : !res;
-      };
-    }
-  }
-  return null;
-}
