+++ lib/util/string.spec.ts
@@ -1,4 +1,4 @@
-import { coerceString, isUUID, looseEquals, replaceAt } from './string';
+import { coerceString, looseEquals, replaceAt } from './string';
 
 describe('util/string', () => {
   describe('replaceAt', () => {
@@ -39,13 +39,5 @@ describe('util/string', () => {
     expect(coerceString(undefined)).toBe('');
     expect(coerceString(null)).toBe('');
     expect(coerceString(null, 'foo')).toBe('foo');
-  });
-
-  describe('isUUID', () => {
-    it('proper checks valid and invalid UUID strings', () => {
-      expect(isUUID('{90b6646d-1724-4a64-9fd9-539515fe94e9}')).toBe(true);
-      expect(isUUID('{90B6646D-1724-4A64-9FD9-539515FE94E9}')).toBe(true);
-      expect(isUUID('not-a-uuid')).toBe(false);
-    });
   });
 });
