+++ lib/workers/repository/process/index.ts
@@ -11,8 +11,8 @@ import { scm } from '../../../modules/platform/scm';
 import { getCache } from '../../../util/cache/repository';
 import { clone } from '../../../util/clone';
 import { getBranchList } from '../../../util/git';
-import { configRegexPredicate } from '../../../util/regex';
 import { addSplit } from '../../../util/split';
+import { configRegexPredicate } from '../../../util/string-match';
 import type { BranchConfig } from '../../types';
 import { readDashboardBody } from '../dependency-dashboard';
 import { ExtractResult, extract, lookup, update } from './extract-update';
